Add function that returns user input if valid
Created a function that takes input from the user, and checks if it's
one of the valid options ("rock", "paper", or "scissors"). If the input
is valid, the value is returned and checked using console.log(),
otherwise an alert dialog is displayed instructing the user to enter a
valid option, then the user is re-prompted to enter input.