Add function to allow playing a set of rounds
I added a function (playGame) that allows you to play a set of rounds.
This is currently set to always play a set of 5.
I moved the round logic, including playRound and calcVictoryState into
it.
The playRound function now tells the user the current score after
every round.
After all rounds have been played, the playGame function tells the user
if they won, lost, or drew.
I also added failsafes to allow the user to cancel the prompt, as
otherwise it would break the script.