Add UI to allow playing without console
Added a paragraph to show the score, a paragraph to show the round
results, and a div that holds three buttons for user input. Once the
score of either the user, or the computer, reaches 5, the winner is
announced, and the buttons are disabled. Can make it replayable in the
future - without refreshing, but the assignment didn't specify that it
had to be replayable.